+++ WalletWasabi.Gui/Tabs/LegalIssuesViewModel.cs
@@ -6,7 +6,7 @@ using WalletWasabi.Gui.ViewModels;
 
 namespace WalletWasabi.Gui.Tabs
 {
-	public class LegalIssuesViewModel : TextViewModelBase
+	public class LegalIssuesViewModel : TextResourceViewModelBase
 	{
 		public LegalIssuesViewModel(Global global) : base(global, "Legal Issues", new Uri("resm:WalletWasabi.Gui.Assets.LegalIssues.txt"))
 		{
