+++ WalletWasabi.Gui/Tabs/PrivacyPolicyViewModel.cs
@@ -6,7 +6,7 @@ using WalletWasabi.Gui.ViewModels;
 
 namespace WalletWasabi.Gui.Tabs
 {
-	public class PrivacyPolicyViewModel : TextViewModelBase
+	public class PrivacyPolicyViewModel : TextResourceViewModelBase
 	{
 		public PrivacyPolicyViewModel(Global global) : base(global, "Privacy Policy", new Uri("resm:WalletWasabi.Gui.Assets.PrivacyPolicy.txt"))
 		{
